requirements.txt
hexbytes==0.1.0
pytest==4.0.1
pytest-forked==0.2
pytest-html==1.19.0
pytest-metadata==1.7.0
pytest-watch==4.2.0
pytest-xdist==1.24.0
requests==2.20.0
web3==4.9.2
PyYAML==4.2b4
paramiko==2.4.2
allure-pytest==2.6.1
ruamel.yaml==0.15.91
pytest-repeat==0.8.0
